+++ Example/Tests/Tests.swift
@@ -1,28 +1,33 @@
+//
+//  Tests.swift
+//  Tests
+//
+//  Created by Tomoya Hirano on 2020/11/13.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
 import XCTest
-import Kitsunebi
 
 class Tests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
+
+    override func setUpWithError() throws {
         // Put setup code here. This method is called before the invocation of each test method in the class.
     }
-    
-    override func tearDown() {
+
+    override func tearDownWithError() throws {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
     }
-    
-    func testExample() {
+
+    func testExample() throws {
         // This is an example of a functional test case.
-        XCTAssert(true, "Pass")
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
     }
-    
-    func testPerformanceExample() {
+
+    func testPerformanceExample() throws {
         // This is an example of a performance test case.
-        self.measure() {
+        measure {
             // Put the code you want to measure the time of here.
         }
     }
-    
+
 }
